refactor(friction_service): extract entry-point + category rollups (#163)
friction_service.py was 602 lines (over the 450 modularity threshold per
issue #163). Extracted friction_entry_points + friction_categories +
their internal helpers (~418 lines) into a new module
friction_entry_points_service.py. Public API preserved via re-export.

  before: friction_service.py = 602 lines
  after:  friction_service.py = 196 lines + friction_entry_points_service.py = 440 lines

321 tests pass.
